Move application list to its own folder
