+++ src/pages/application-inventory/application-list/components/application-form/application-form.tsx
