+++ src/pages/application-inventory/application-list/components/application-form/index.ts
