+++ src/pages/application-inventory/application-list/components/application-form/stories/application-form.stories.tsx
