+++ src/pages/application-inventory/application-list/components/new-application-modal/index.ts
