+++ src/pages/application-inventory/application-list/components/new-application-modal/new-application-modal.tsx
