+++ src/pages/application-inventory/application-list/components/remote-assessment/index.ts
