+++ src/pages/application-inventory/application-list/components/remote-assessment/remote-assessment.tsx
