+++ src/pages/application-inventory/application-list/components/remote-business-service/index.ts
