+++ src/pages/application-inventory/application-list/components/remote-business-service/remote-business-service.tsx
