+++ src/pages/application-inventory/application-list/components/update-application-modal/index.ts
