+++ src/pages/application-inventory/application-list/index.ts
@@ -1,0 +1,1 @@
+export { ApplicationList as default } from "./application-inventory";
